+++ src/routes/AppRouter.js
@@ -1,0 +1,41 @@
+import { RouterProvider, createBrowserRouter } from "react-router-dom";
+import Main from "../pages/mainPage/Main";
+import PreHeader from "../pages/headers/PreHeader";
+import Login from "../pages/loginPage/Login";
+import Signup from "../pages/signUpPage/Signup";
+import Error from "../pages/errorPage/Error";
+import PostHeader from "../pages/headers/PostHeader";
+
+const AppRouter = () => {
+  const router = createBrowserRouter([
+    {
+      path: "/",
+      element: <PreHeader />,
+      errorElement: <Error />,
+      children: [
+        {
+          path: "",
+          element: <Main />,
+        },
+        {
+          path: "login",
+          element: <Login />,
+        },
+        {
+          path: "signup",
+          element: <Signup />,
+        },
+      ],
+    },
+    {
+      path: "/study",
+      element: <PostHeader />,
+      errorElement: <Error />,
+      children: [{}],
+    },
+  ]);
+
+  return <RouterProvider router={router} />;
+};
+
+export default AppRouter;
